quang

+++ LIBPRO v2.0 full op/Source.cpp
@@ -224,7 +224,7 @@ int TruyXuatThongBao(ThongBao v[]) {
 	return dem - 1;
 }
 void CapNhatNgDung(NguoiDung v[]) {
-	int m;
+	double m;
 	fstream outFile("UserData.DAT", ios::app);
 	do {
 		cin.clear();
